editor/libeditor/tests/test_password_unmask_API.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for unmasking password API</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<input type="text">
<input type="password">
<script class="testbody" type="application/javascript">
SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(() => {
  let input = document.getElementsByTagName("input")[0];
  let password = document.getElementsByTagName("input")[1];

  let editor = SpecialPowers.wrap(input).editor;
  let passwordEditor = SpecialPowers.wrap(password).editor;

  try {
    editor.mask();
    ok(false,
       `nsIEditor.mask() should throw exception when called for <input type="text"> before nsIEditor.unmask()`);
  } catch (e) {
    ok(true,
       `nsIEditor.mask() should throw exception when called for <input type="text"> before nsIEditor.unmask() ${e}`);
  }

  try {
    editor.unmask();
    ok(false,
       `nsIEditor.unmask() should throw exception when called for <input type="text">`);
  } catch (e) {
    ok(true,
       `nsIEditor.unmask() should throw exception when called for <input type="text"> ${e}`);
  }

  try {
    editor.unmask(0);
    ok(false,
       `nsIEditor.unmask(0) should throw exception when called for <input type="text">`);
  } catch (e) {
    ok(true,
       `nsIEditor.unmask(0) should throw exception when called for <input type="text"> ${e}`);
  }

  input.value = "abcdef";
  try {
    editor.unmask();
    ok(false,
       `nsIEditor.unmask() should throw exception when called for <input type="text" value="abcdef">`);
  } catch (e) {
    ok(true,
       `nsIEditor.unmask() should throw exception when called for <input type="text" value="abcdef"> ${e}`);
  }

  try {
    editor.mask();
    ok(false,
       `nsIEditor.mask() should throw exception when called for <input type="text" value="abcdef"> after nsIEditor.unmask()`);
  } catch (e) {
    ok(true,
       `nsIEditor.mask() should throw exception when called for <input type="text" value="abcdef"> after nsIEditor.unmask() ${e}`);
  }

  try {
    passwordEditor.mask();
    ok(true,
       `nsIEditor.mask() shouldn't throw exception when called for <input type="password"> before nsIEditor.unmask()`);
    ok(passwordEditor.autoMaskingEnabled,
       `nsIEditor.autoMaskingEnabled should be true after nsIEditor.mask() for <input type="password"> before nsIEditor.unmask()`);
  } catch (e) {
    ok(false,
       `nsIEditor.mask() shouldn't throw exception when called for <input type="password"> before nsIEditor.unmask() ${e}`);
  }

  try {
    editor.unmask(5);
    ok(false,
       `nsIEditor.unmask(5) should throw exception when called for <input type="password" value="">`);
  } catch (e) {
    ok(true,
       `nsIEditor.unmask(5) should throw exception when called for <input type="password" value=""> ${e}`);
    ok(passwordEditor.autoMaskingEnabled,
       `nsIEditor.autoMaskingEnabled should keep true (<input type="password">)`);
  }

  try {
    passwordEditor.unmask();
    ok(true,
       `nsIEditor.unmask() shouldn't throw exception when called for <input type="password">`);
    ok(!passwordEditor.autoMaskingEnabled,
       `nsIEditor.autoMaskingEnabled should be false after nsIEditor.unmask() for <input type="password">)`);
    is(passwordEditor.unmaskedStart, 0,
       `nsIEditor.unmaskedStart should be 0 after nsIEditor.unmask() for <input type="password">`);
    is(passwordEditor.unmaskedEnd, 0,
       `nsIEditor.unmaskedEnd should be 0 after nsIEditor.unmask() for <input type="password">`);
  } catch (e) {
    ok(false,
       `nsIEditor.unmask() shouldn't throw exception when called for <input type="password"> ${e}`);
  }

  password.value = "abcdef";
  try {
    passwordEditor.unmask();
    ok(true,
       `nsIEditor.unmask() shouldn't throw exception when called for <input type="password" value="abcdef">)`);
    ok(!passwordEditor.autoMaskingEnabled,
       `nsIEditor.autoMaskingEnabled should be false after nsIEditor.unmask() for <input type="password" value="abcdef">`);
    is(passwordEditor.unmaskedStart, 0,
       `nsIEditor.unmaskedStart should be 0 after nsIEditor.unmask() for <input type="password" value="abcdef">`);
    is(passwordEditor.unmaskedEnd, 6,
       `nsIEditor.unmaskedEnd should be 0 after nsIEditor.unmask() for <input type="password" value="abcdef">`);
  } catch (e) {
    ok(false,
       `nsIEditor.unmask() shouldn't throw exception when called for <input type="password" value="abcdef"> ${e}`);
  }

  try {
    passwordEditor.mask();
    ok(true,
       `nsIEditor.mask() shouldn't throw exception when called for <input type="password" value="abcdef">`);
    ok(passwordEditor.autoMaskingEnabled,
       `nsIEditor.autoMaskingEnabled should be true after nsIEditor.mask() for <input type="password" value="abcdef">`);
  } catch (e) {
    ok(false,
       `nsIEditor.mask() shouldn't throw exception when called for <input type="password" value="abcdef"> ${e}`);
  }

  try {
    passwordEditor.unmask(0, 100, 1000);
    ok(true,
       `nsIEditor.unmask(0, 100, 1000) shouldn't throw exception when called for <input type="password" value="abcdef">`);
    ok(passwordEditor.autoMaskingEnabled,
       `nsIEditor.autoMaskingEnabled should be true after nsIEditor.unmask(0, 100, 1000) for <input type="password" value="abcdef">`);
    is(passwordEditor.unmaskedStart, 0,
       `nsIEditor.unmaskedStart should be 0 after nsIEditor.unmask(0, 100, 1000) for <input type="password" value="abcdef">`);
    is(passwordEditor.unmaskedEnd, 6,
       `nsIEditor.unmaskedEnd should be 6 after nsIEditor.unmask(0, 100, 1000) for <input type="password" value="abcdef">`);
  } catch (e) {
    ok(false,
       `nsIEditor.unmask(0, 100, 1000) shouldn't throw exception when called for <input type="password" value="abcdef"> ${e}`);
  }

  try {
    passwordEditor.unmask(3);
    ok(true,
       `nsIEditor.unmask(3) shouldn't throw exception when called for <input type="password" value="abcdef">`);
    ok(!passwordEditor.autoMaskingEnabled,
       `nsIEditor.autoMaskingEnabled should be false after nsIEditor.unmask(3) for <input type="password" value="abcdef">`);
    is(passwordEditor.unmaskedStart, 3,
       `nsIEditor.unmaskedStart should be 3 after nsIEditor.unmask(3) for <input type="password" value="abcdef">`);
    is(passwordEditor.unmaskedEnd, 6,
       `nsIEditor.unmaskedEnd should be 6 after nsIEditor.unmask(3) for <input type="password" value="abcdef">`);
  } catch (e) {
    ok(false,
       `nsIEditor.unmask(3) shouldn't throw exception when called for <input type="password" value="abcdef"> ${e}`);
  }

  SimpleTest.finish();
});
</script>
</body>
</html>
